QS: Only reduce row count if media player is visible -- Genkzsz11: Also remove import static
There's no need to always force qs_less_rows here. The QS layout code
will automatically reduce the amount of rows in order to fit everything
on the screen when the media player is open, so we can still use the
extra space when the media player is not in use.

Change-Id: I99101c6816b402a87d9e429cdcd0be86874bb26b