gmscompat: fix push notifications in secondary user profiles
UserManager.isUserUnlocked(userId) returns whether a user exited the Direct Boot mode.
This method requires a restricted permission if userId parameter differs from the current userId.
GMS called this method with userId 0 in both primary (userId == 0) and secondary (userId != 0)
profiles. userId was always 0 likely because of the stubbed out
UserManager.getUserSerialNumber(userId),
which always returned 0 as a workaround to a different issue.

Signed-off-by: Dmitrii <bankersenator@gmail.com>